Configured mbed os board as VMN coap server
Server is open forever to listen for client requests. Here, client keep on reattempts packets until it receives response from the server within the response timeout. If server doesn't respond within the response time client stops sending packets.  At the moment server is freezing because of the continuous loop, still working on this to resolve the issue.